+++ src/ical/client.rs
@@ -7,14 +7,10 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use chrono::{Duration, NaiveDate, NaiveDateTime};
 use ical::{
     IcalParser,
     parser::ical::component::{IcalEvent, IcalTodo},
-    property::Property,
 };
-
-use crate::task::DateTimeUtc;
 
 use super::task::Task;
 
@@ -87,95 +83,12 @@ where
     Ok(tasks)
 }
 
-fn tz_offset_from_property_params(params: &Option<Vec<(String, Vec<String>)>>) -> Option<chrono_tz::Tz> {
-    if let Some(params) = params {
-        for (n, p) in params {
-            if n == "TZID"
-                && p.len() == 1
-                && let Ok(t) = p[0].parse::<chrono_tz::Tz>()
-            {
-                return Some(t);
-            }
-        }
-    }
-
-    None
-}
-
-fn dt_from_property(p: &Property) -> Option<DateTimeUtc> {
-    let s = p.value.as_ref()?;
-
-    if let Ok(d) = NaiveDate::parse_from_str(s, "%Y%m%d") {
-        let dt = d.and_hms_opt(0, 0, 0)?;
-        return Some(DateTimeUtc::from_naive_utc_and_offset(dt, chrono::Utc));
-    }
-
-    // with timezone in params
-    if let Some(tz) = tz_offset_from_property_params(&p.params)
-        && let Ok(dt) = NaiveDateTime::parse_from_str(s, "%Y%m%dT%H%M%S")
-    {
-        let dt = dt.and_local_timezone(tz).unwrap();
-        return Some(dt.to_utc());
-    }
-
-    // with timezone inside
-    if let Ok(dt) = NaiveDateTime::parse_from_str(s, "%Y%m%dT%H%M%SZ") {
-        return Some(DateTimeUtc::from_naive_utc_and_offset(dt, chrono::Utc));
-    }
-
-    None
-}
-
-fn duration_from_property(p: &Property) -> Option<Duration> {
-    if let Some(v) = &p.value
-        && let Ok(d) = v.parse::<iso8601_duration::Duration>()
-    {
-        return d.to_chrono();
-    }
-
-    None
-}
-
-fn fill_task(t: &mut Task, properties: &[Property]) {
-    t.properties = properties.to_vec();
-
-    for p in properties {
-        match p.name.as_str() {
-            "UID" => t.uid = p.value.clone().unwrap_or_default(),
-            "SUMMARY" => t.name = p.value.clone().unwrap_or_default(),
-            "DESCRIPTION" => t.description = p.value.clone(),
-            "PRIORITY" => t.priority = p.value.as_ref().map(|s| s.parse::<u8>().unwrap_or(0)).unwrap_or(0),
-            "DUE" => t.due = dt_from_property(p),
-            "DTSTART" => t.start = dt_from_property(p),
-            "DTEND" => t.end = dt_from_property(p),
-            "COMPLETED" => t.completed = dt_from_property(p),
-            "CREATED" => t.created = dt_from_property(p),
-            "DURATION" => t.duration = duration_from_property(p),
-            "CATEGORIES" if p.value.is_some() => {
-                t.categories = p
-                    .value
-                    .as_ref()
-                    .unwrap()
-                    .split(",")
-                    .map(|s| s.to_string())
-                    .collect::<Vec<String>>()
-            }
-            _ => {}
-        }
-    }
-    tracing::debug!(task=?t, "New task");
-}
-
 fn event_to_task(ev: &IcalEvent) -> Task {
-    let mut t = Task::default();
-    fill_task(&mut t, &ev.properties);
-    t
+    Task::from(&ev.properties)
 }
 
 fn todo_to_task(todo: &IcalTodo) -> Task {
-    let mut t = Task::default();
-    fill_task(&mut t, &todo.properties);
-    t
+    Task::from(&todo.properties)
 }
 
 #[cfg(test)]
